src/main/java/com/app/urlshortner/model/URLShortenRequest.java
package com.app.urlshortner.model;

import javax.validation.constraints.Min;
import javax.validation.constraints.NotNull;

import com.fasterxml.jackson.annotation.JsonProperty;

import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;

@ApiModel(value = "URLShortenRequest")
public class URLShortenRequest {

    @ApiModelProperty
    @JsonProperty("url")
    @NotNull
    private String url;

    public URLShortenRequest() {}

    public URLShortenRequest(String inputURL) {
        this.url = inputURL;
    }

    public String getURL() {
        return url;
    }

    public void setURL(String URL) {
        this.url = URL;
    }

    @Override
    public int hashCode() {
        final int prime = 31;
        int result = 1;
        result = prime * result + ((url == null) ? 0 : url.hashCode());
        return result;
    }

    @Override
    public boolean equals(Object obj) {
        if (this == obj)
            return true;
        if (obj == null)
            return false;
        if (getClass() != obj.getClass())
            return false;
        URLShortenRequest other = (URLShortenRequest) obj;
        if (url == null) {
            if (other.url != null)
                return false;
        } else if (!url.equals(other.url))
            return false;
        return true;
    }

}
